chore: refactor mineru-web-api
