+++ mineru_web_api/app/app.py
@@ -1,10 +1,10 @@
 import os
 import tempfile
-from io import BytesIO, StringIO
-from typing import Annotated, Tuple
+from typing import Annotated
 
 import magic_pdf.model as model_config
 import uvicorn
+from data_reader_writer import MemoryDataWriter, image_writer, markdown_writer
 from fastapi import FastAPI, File, Form, UploadFile
 from fastapi.responses import JSONResponse
 from langchain_core.documents import Document
@@ -17,130 +17,15 @@ from magic_pdf.data.read_api import read_local_images, read_local_office
 from magic_pdf.model.doc_analyze_by_custom_model import doc_analyze
 from magic_pdf.operators.models import InferenceResult
 from magic_pdf.operators.pipes import PipeResult
-from minio import Minio
 from vector_store import vector_store
 
 model_config.__use_inside_model__ = True
 
 app = FastAPI()
 
-# MinIO configuration
-MINIO_URL = os.environ.get("MINIO_ENDPOINT", "localhost:9000")
-IS_PROD = os.getenv("ENVIRONMENT") == "production"
-
-if IS_PROD:
-    MINIO_URL = "minio.cspc.me"
-
-MINIO_ACCESS_KEY = os.environ.get("MINIO_ACCESS_KEY", "minioadmin")
-MINIO_SECRET_KEY = os.environ.get("MINIO_SECRET_KEY", "minioadmin")
-
-minio_client = Minio(
-    MINIO_URL,
-    access_key=MINIO_ACCESS_KEY,
-    secret_key=MINIO_SECRET_KEY,
-    secure=IS_PROD,
-)
-
-
-# Ensure buckets exist
-def ensure_bucket_exists(bucket_name: str):
-    if not minio_client.bucket_exists(bucket_name):
-        minio_client.make_bucket(bucket_name)
-
-
 pdf_extensions = [".pdf"]
 office_extensions = [".ppt", ".pptx", ".doc", ".docx"]
 image_extensions = [".png", ".jpg", ".jpeg"]
-
-
-class MemoryDataWriter(DataWriter):
-    def __init__(self):
-        self.buffer = StringIO()
-
-    def write(self, path: str, data: bytes) -> None:
-        if isinstance(data, str):
-            self.buffer.write(data)
-        else:
-            self.buffer.write(data.decode("utf-8"))
-
-    def write_string(self, path: str, data: str) -> None:
-        self.buffer.write(data)
-
-    def get_value(self) -> str:
-        return self.buffer.getvalue()
-
-    def close(self):
-        self.buffer.close()
-
-
-class MinioDataWriter(DataWriter):
-    """Data writer implementation for MinIO storage"""
-
-    def __init__(self, bucket_name: str):
-        self.bucket_name = bucket_name
-        ensure_bucket_exists(bucket_name)
-
-    def write(self, path: str, data: bytes) -> None:
-        """
-        Write data to MinIO storage
-        Args:
-            path: Object name in the bucket
-            data: Data to be written
-        """
-        if isinstance(data, str):
-            data_bytes = data.encode("utf-8")
-        else:
-            data_bytes = data
-
-        minio_client.put_object(
-            bucket_name=self.bucket_name,
-            object_name=path,
-            data=BytesIO(data_bytes),
-            length=len(data_bytes),
-        )
-
-    def write_string(self, path: str, data: str) -> None:
-        """
-        Write string data to MinIO storage
-        Args:
-            path: Object name in the bucket
-            data: String data to be written
-        """
-        data_bytes = data.encode("utf-8")
-        minio_client.put_object(
-            bucket_name=self.bucket_name,
-            object_name=path,
-            data=BytesIO(data_bytes),
-            length=len(data_bytes),
-            content_type="text/plain",
-        )
-
-
-def init_writers(
-    file: UploadFile,
-    project_id: str,
-) -> Tuple[DataWriter, DataWriter, bytes, str]:
-    """
-    Initialize writers based on path type
-
-    Args:
-        file: Uploaded file object
-        project_id: Project ID of the file
-
-    Returns:
-        Tuple[md_writer, image_writer, file_bytes, file_extension]
-    """
-    # Read the uploaded file content
-    file_bytes = file.file.read()
-    if file.filename is None:
-        raise ValueError("File name is missing.")
-    file_extension = os.path.splitext(file.filename)[1].lower()
-
-    # Create writers for "markdowns" and "images" buckets
-    markdown_writer = MinioDataWriter("markdowns")
-    image_writer = MinioDataWriter("images")
-
-    return markdown_writer, image_writer, file_bytes, file_extension
 
 
 def process_file(
@@ -242,7 +127,8 @@ async def file_parse(
 ):
     """
     Execute the process of converting a document file to markdown content.
-    Also saves the content to vector store for later retrieval.
+    Save the markdown content and image files to object storage.
+    Split the markdown content into documents and add them to the vector store.
 
     Args:
         file: The file to be parsed (PDF, Office document, or image)
@@ -252,15 +138,12 @@ async def file_parse(
         return JSONResponse({"error": "File name is missing."}, status_code=400)
 
     try:
-        # Initialize readers/writers and get file content
-        md_writer, image_writer, file_bytes, file_extension = init_writers(
-            file=file, project_id=project_id
-        )
+        # Get the file content and extension
+        file_bytes = file.file.read()
+        file_extension = os.path.splitext(file.filename)[1].lower()
 
         # Process the file
-        infer_result, pipe_result = process_file(
-            file_bytes, file_extension, image_writer
-        )
+        _, pipe_result = process_file(file_bytes, file_extension, image_writer)
 
         # Use MemoryDataWriter to get results
         md_content_writer = MemoryDataWriter()
@@ -274,7 +157,7 @@ async def file_parse(
         # Save markdown content to MinIO
         file_name_without_ext = os.path.splitext(file.filename)[0]
         markdown_path = f"{project_id}/{file_name_without_ext}.md"
-        md_writer.write_string(markdown_path, md_content)
+        markdown_writer.write_string(markdown_path, md_content)
 
         # Clean up memory writers
         md_content_writer.close()
